Splits the browser initialization steps in their own file - works fine. Will split other tests as needed.
